+++ instant_import/static/src/js/instant_import.js
@@ -1,9 +1,5 @@
 /** @odoo-module **/
-
-import { DropdownItem } from "@web/core/dropdown/dropdown_item";
-import { registry } from "@web/core/registry";
 import { useService } from "@web/core/utils/hooks";
-import { Component } from "@odoo/owl";
 import { patch } from "@web/core/utils/patch";
 import { ImportRecords } from "@base_import/import_records/import_records";