+++ instant_import/doc/RELEASE_NOTES.md
@@ -1,7 +1,7 @@
-## Module <custom_import>
+## Module <instant_import>
 
-#### 17.07.2025
+#### 17.09.2025
 #### Version 17.0.1.0.0
 ##### ADD
 
-- Initial Commit for Custom Import
+- Initial Commit for Instant Import